+++ src/assets/images/index.tsx
@@ -2,3 +2,4 @@ export * from './IconEggNoPlan';
 export * from './IconEggOnePlan';
 export * from './IconEggThreePlan';
 export * from './IconEggTwoPlan';
+export * from './Images';
