--- src/components/common/Images/index.tsx
@@ -1,1 +1,0 @@
-export { Images } from './Images'
